+++ buildSrc/src/main/kotlin/com/orange/ods/gradle/Dependencies.kt
@@ -13,6 +13,8 @@ package com.orange.ods.gradle
 object Dependencies {
 
     const val accompanistSystemUiController = "com.google.accompanist:accompanist-systemuicontroller:${Versions.accompanist}"
+    const val accompanistPager = "com.google.accompanist:accompanist-pager:${Versions.accompanist}"
+    const val accompanistPagerIndicators = "com.google.accompanist:accompanist-pager-indicators:${Versions.accompanist}"
     const val activityCompose = "androidx.activity:activity-compose:${Versions.activity}"
     const val androidGradlePlugin = "com.android.tools.build:gradle:${Versions.androidGradlePlugin}"
     const val appCompat = "androidx.appcompat:appcompat:${Versions.appCompat}"