+++ demo/src/main/java/com/orange/ods/demo/ui/MainActivity.kt
@@ -14,8 +14,10 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.material.ExperimentalMaterialApi
+import com.google.accompanist.pager.ExperimentalPagerApi
 
 @ExperimentalMaterialApi // Used for BottomSheetScaffold in the ComponentsCardScreen
+@ExperimentalPagerApi
 class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {