+++ demo/src/main/java/com/orange/ods/demo/ui/OdsDemoApp.kt
@@ -17,10 +17,13 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material.AppBarDefaults
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Icon
 import androidx.compose.material.Scaffold
+import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -33,14 +36,19 @@ import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.NavHost
 import androidx.navigation.navigation
+import com.google.accompanist.pager.ExperimentalPagerApi
 import com.orange.ods.compose.component.bottomnavigation.OdsBottomNavigation
 import com.orange.ods.compose.component.bottomnavigation.OdsBottomNavigationItem
 import com.orange.ods.compose.theme.OdsMaterialTheme
 import com.orange.ods.demo.ui.about.addAboutGraph
 import com.orange.ods.demo.ui.components.addComponentsGraph
+import com.orange.ods.demo.ui.components.tabs.TabsConfiguration
+import com.orange.ods.demo.ui.components.tabs.TopAppBarFixedTabs
+import com.orange.ods.demo.ui.components.tabs.TopAppBarScrollableTabs
 import com.orange.ods.demo.ui.guidelines.addGuidelinesGraph
 
 @ExperimentalMaterialApi
+@ExperimentalPagerApi
 @Preview(showBackground = true, uiMode = UI_MODE_NIGHT_YES)
 @Preview(showBackground = true)
 @Composable
@@ -58,13 +66,40 @@ fun OdsDemoApp() {
     OdsMaterialTheme(appState.darkModeEnabled.value) {
         Scaffold(
             topBar = {
-                OdsDemoTopAppBar(
-                    titleRes = appState.topAppBarTitleRes.value,
-                    darkModeEnabled = appState.darkModeEnabled.value,
-                    shouldShowUpNavigationIcon = !appState.shouldShowBottomBar,
-                    navigateUp = appState::upPress,
-                    updateTheme = appState::updateTheme
-                )
+                Surface(elevation = AppBarDefaults.TopAppBarElevation) {
+                    Column {
+                        OdsDemoTopAppBar(
+                            titleRes = appState.topAppBarTitleRes.value,
+                            darkModeEnabled = appState.darkModeEnabled.value,
+                            shouldShowUpNavigationIcon = !appState.shouldShowBottomBar,
+                            navigateUp = appState::upPress,
+                            updateTheme = appState::updateTheme
+                        )
+                        // Display tabs in the top bar if needed
+                        with(appState.tabsState) {
+                            pagerState?.let { pagerState ->
+                                if (hasTabs) {
+                                    if (scrollableTabs.value) {
+                                        TopAppBarScrollableTabs(
+                                            tabs = tabs,
+                                            pagerState = pagerState,
+                                            tabIconType = tabIconType.value,
+                                            tabTextEnabled = tabTextEnabled.value
+                                        )
+                                    } else {
+                                        TopAppBarFixedTabs(
+                                            tabs = tabs,
+                                            pagerState = pagerState,
+                                            tabIconType = tabIconType.value,
+                                            tabTextEnabled = tabTextEnabled.value
+                                        )
+                                    }
+                                }
+                            }
+                        }
+
+                    }
+                }
             },
             bottomBar = {
                 AnimatedVisibility(
@@ -84,7 +119,9 @@ fun OdsDemoApp() {
                 NavHost(appState.navController, startDestination = MainDestinations.HOME_ROUTE) {
                     odsDemoNavGraph(
                         onNavElementClick = appState::navigateToElement,
-                        updateTopBarTitle = appState::updateTopAppBarTitle
+                        updateTopBarTitle = appState::updateTopAppBarTitle,
+                        updateTopAppBarTabs = appState.tabsState::updateTopAppBarTabs,
+                        clearTopAppBarTabs = appState.tabsState::clearTopAppBarTabs
                     )
                 }
             }
@@ -106,19 +143,22 @@ private fun OdsDemoBottomBar(tabs: Array<HomeSections>, currentRoute: String, na
     }
 }
 
+@ExperimentalPagerApi
 @ExperimentalMaterialApi
 private fun NavGraphBuilder.odsDemoNavGraph(
     onNavElementClick: (String, Long?, NavBackStackEntry) -> Unit,
-    updateTopBarTitle: (Int) -> Unit
+    updateTopBarTitle: (Int) -> Unit,
+    updateTopAppBarTabs: (TabsConfiguration) -> Unit,
+    clearTopAppBarTabs: () -> Unit
 ) {
     navigation(
         route = MainDestinations.HOME_ROUTE,
         startDestination = HomeSections.GUIDELINES.route
     ) {
-        addHomeGraph(onNavElementClick, updateTopBarTitle)
+        addHomeGraph(onNavElementClick, updateTopBarTitle, clearTopAppBarTabs)
     }
 
-    addGuidelinesGraph(updateTopBarTitle)
-    addComponentsGraph(onNavElementClick, updateTopBarTitle)
-    addAboutGraph(updateTopBarTitle)
+    addGuidelinesGraph(updateTopBarTitle, clearTopAppBarTabs)
+    addComponentsGraph(onNavElementClick, updateTopBarTitle, updateTopAppBarTabs, clearTopAppBarTabs)
+    addAboutGraph(updateTopBarTitle, clearTopAppBarTabs)
 }