+++ lib/src/main/java/com/orange/ods/compose/component/tab/OdsTabRow.kt
@@ -1,0 +1,53 @@
+/*
+ *
+ *  Copyright 2021 Orange
+ *
+ *  Use of this source code is governed by an MIT-style
+ *  license that can be found in the LICENSE file or at
+ *  https://opensource.org/licenses/MIT.
+ * /
+ */
+
+package com.orange.ods.compose.component.tab
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Tab
+import androidx.compose.material.TabRow
+import androidx.compose.material.TabRowDefaults
+import androidx.compose.material.TabRowDefaults.tabIndicatorOffset
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+/**
+ * <a href="https://system.design.orange.com/0c1af118d/p/513d27-tabs/b/50cb71" class="external" target="_blank">ODS tabs</a>.
+ *
+ * An OdsTabRow is a Jetpack Compose [TabRow] to which we applied the Orange design and theme.
+ * @see TabRow documentation
+ *
+ * @param selectedTabIndex the index of the currently selected tab
+ * @param modifier optional [Modifier] for this OdsTabRow
+ * @param tabs the tabs inside this TabRow. Typically this will be multiple [Tab]s. Each element
+ * inside this lambda will be measured and placed evenly across the TabRow, each taking up equal
+ * space. Use [OdsTab] to display Orange styled tabs.
+ */
+@Composable
+fun OdsTabRow(
+    selectedTabIndex: Int,
+    modifier: Modifier = Modifier,
+    tabs: @Composable () -> Unit
+) {
+    TabRow(
+        modifier = modifier,
+        backgroundColor = MaterialTheme.colors.surface,
+        contentColor = MaterialTheme.colors.onSurface,
+        selectedTabIndex = selectedTabIndex,
+        indicator = { tabPositions ->
+            TabRowDefaults.Indicator(
+                modifier = Modifier.tabIndicatorOffset(tabPositions[selectedTabIndex]),
+                color = MaterialTheme.colors.primary
+            )
+        },
+        divider = {},
+        tabs = tabs
+    )
+}